Task #463 creating of a new class EListStringCellEditingSupport
to handle the adding of new user in the Discipline.
